feat: 유저 role 체킹 개발
